requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.0.0
appnope==0.1.3
arrow==1.2.3
asttokens==2.4.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
azure-cognitiveservices-speech==1.28.0
backcall==0.2.0
binaryornot==0.4.4
botbuilder-core==4.14.4
botbuilder-integration-aiohttp==4.14.4
botbuilder-schema==4.14.4
botframework-connector==4.14.4
botframework-streaming==4.14.4
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
cookiecutter==1.7.0
cryptography==41.0.3
debugpy==1.7.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
frozenlist==1.4.0
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
isodate==0.6.1
jedi==0.19.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpickle==1.4.2
jupyter_client==8.3.1
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msal==1.23.0
msrest==0.6.21
multidict==6.0.4
nest-asyncio==1.5.7
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
poyo==0.5.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.5
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
whichcraft==0.6.1
yarl==1.4.2
zipp==3.16.2
